Added a response_id field to the JWT
This field is used to uniquely identify a set of questionnaire
responses. It differs from the case identifier in the following way:
in the event there are multiple responses for a case (e.g. on account
of multiple issued UACs), the case identifier will be the same but
response_id will differ